+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [js/v0.2.10](https://github.com/runapi-ai/suno-sdk/releases/tag/js%2Fv0.2.10), [ruby/v0.2.10](https://github.com/runapi-ai/suno-sdk/releases/tag/ruby%2Fv0.2.10), [go/v0.2.10](https://github.com/runapi-ai/suno-sdk/releases/tag/go%2Fv0.2.10), [python/v0.1.3](https://github.com/runapi-ai/suno-sdk/releases/tag/python%2Fv0.1.3) - 2026-07-20
+
+### Changed
+- Validate that Suno mashup requests contain exactly 2 upload URLs.
+
+
 ## [js/v0.2.9](https://github.com/runapi-ai/suno-sdk/releases/tag/js%2Fv0.2.9), [ruby/v0.2.9](https://github.com/runapi-ai/suno-sdk/releases/tag/ruby%2Fv0.2.9), [go/v0.2.9](https://github.com/runapi-ai/suno-sdk/releases/tag/go%2Fv0.2.9), [python/v0.1.2](https://github.com/runapi-ai/suno-sdk/releases/tag/python%2Fv0.1.2), [java/v0.1.3](https://github.com/runapi-ai/suno-sdk/releases/tag/java%2Fv0.1.3) - 2026-07-20
 
 ### Added